+++ knowledge/pages/general/CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [0.3.0] - 2024-03-15
 
+### Changed
+
+- Whitespace at the end of the body is no longer removed during decoding.
+
 ### Removed
 
 - Support for PHP < 8